refactor(api): move Prescriber model
to team bounded context
