+++ api/src/team/infrastructure/repositories/prescriber-repository.js
@@ -6,7 +6,7 @@ import { Tag } from '../../../../lib/domain/models/Tag.js';
 import { UserOrgaSettings } from '../../../../lib/domain/models/UserOrgaSettings.js';
 import { Organization } from '../../../organizational-entities/domain/models/Organization.js';
 import { ForbiddenAccess } from '../../../shared/domain/errors.js';
-import { Prescriber } from '../../../shared/prescriber-management/domain/read-models/Prescriber.js';
+import { Prescriber } from '../../domain/read-models/Prescriber.js';
 
 function _toPrescriberDomain(user, userOrgaSettings, tags, memberships, organizations, schools) {
   return new Prescriber({
